create custom _document.tsx
By default, Next.js abstracts away the `<html>` and `<body>` tags on your pages.

To use Material UI, your `theme` and `createEmotionCache` instances need to be able to talk to these tags. This happens in the custom `_document` file.

The custom `_document` also lets you set the `initialProps` for the page, so styles can be rendered on the server before the page is sent to the client.